requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
requests>=2.31.0
pillow>=10.0.0
slack-sdk>=3.23.0
